requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
requests
pandas
geopy
folium
pytest
